# client/public/openapi.yaml
openapi: 3.0.3
info:
  title: AstraCalls API
  version: "1.0.0"
  description: |
    API HTTP do **AstraCalls** — chamadas de voz do WhatsApp pelo navegador.

    ## Autenticação
    Não há login/senha. As rotas `/api/*` são protegidas por API key: envie o header
    **`X-API-Key`** com o valor definido em `WACALLS_API_KEY` (ou `?apiKey=...` na query).
    Se `WACALLS_API_KEY` estiver vazia, a autenticação fica desligada.
    Exceção de segurança: downloads de gravações e o arquivo administrativo exigem
    uma chave mestra não vazia no header `X-API-Key`; não aceitam chave na URL ou
    chave de widget, mesmo quando outras rotas permitem esses mecanismos.

    O header **`X-Client-Id`** é opcional e **não autentica nada** — é só um identificador
    estável usado para marcar o "dono" das chamadas e filtrar eventos.

    > ⚠️ A API inicia/controla as chamadas (sinalização), mas o **áudio** exige um cliente
    > WebRTC (navegador) com captura de microfone — não é possível só com chamadas HTTP.
servers:
  - url: /
    description: Mesma origem do painel
...
security:
  - apiKeyAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key definida em WACALLS_API_KEY (vazia = sem autenticação)"
  parameters:
    ClientId:
      name: X-Client-Id
      in: header
      required: false
      schema: { type: string, example: meuapp }
      description: >-
        Opcional. Identificador do cliente (dono da chamada / filtro de eventos).
        NÃO é autenticação — para autenticar use o header X-API-Key.
    Sid:
      name: sid
      in: path
      required: true
      schema: { type: string }
      description: ID da sessão (conta WhatsApp).
...
  schemas:
    Session:
      type: object
      properties:
        id: { type: string, example: 6458cca60a17c0cefa71de9da9dc92a2 }
        name: { type: string, example: WhatsApp }
        jid: { type: string, example: "556192264108:13@s.whatsapp.net" }
        state: { type: string, enum: [open, connecting, close], example: open }
        paired: { type: boolean, example: true }
        recording: { type: boolean, example: false, description: "Gravação de chamadas ligada para esta conta" }
...
paths:
...
  /api/sessions:
    get:
      tags: [Sessões]
      summary: Lista as contas
      parameters: [{ $ref: "#/components/parameters/ClientId" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
    post:
      tags: [Sessões]
      summary: Cria uma conta
      parameters: [{ $ref: "#/components/parameters/ClientId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: "Minha conta" }
      responses:
        "200":
          description: Criada
          content:
            application/json:
              schema:
                type: object
                properties: { id: { type: string } }
  /api/sessions/{sid}/recording:
    get:
      tags: [Sessões]
      summary: Estado da gravação de chamadas da conta
      parameters:
        - { $ref: "#/components/parameters/Sid" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { enabled: { type: boolean } }
    put:
      tags: [Sessões]
      summary: Liga/desliga a gravação de chamadas da conta
      description: |
        Ativa a captura nas chamadas QR suportadas desta conta. Chamadas abaixo
        de 3 segundos não geram arquivo; a captura tem limite de 60 minutos.
        O MP3 mistura a voz original recebida do interlocutor com o áudio enviado
        a ele (que pode estar traduzido); não contém necessariamente todas as
        vozes originais e traduções. Gravação de chamadas Meta oficiais não é suportada.
        No fim, o MP3 pode ser enviado como nota privada ao Chatwoot configurado,
        e o evento `recording` inclui `requiresMasterHeader: true`.
        `GET /recordings/{callId}.mp3` exige a chave mestra no header `X-API-Key`.
        O arquivo de conversas e sua exclusão automática são opções separadas,
        desativadas por padrão. Sua política configurável propõe originais por
        24 horas e resumos por 90 dias após o fechamento manual, suspendendo a
        exclusão durante reclamações ou retenções legais. Gravações antigas não
        indexadas são preservadas; não há mais limpeza indiscriminada de 48 horas.
        O arquivo aceita chamadas até 64 MiB e mensagens de voz até 25 MiB.
        Cópias externas e backups exigem políticas de retenção independentes.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties: { enabled: { type: boolean, example: true } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { enabled: { type: boolean } }
...
  /recordings/{id}:
    get:
      tags: [Sessões]
      summary: Baixa um MP3 privado com autenticação mestra no header
      description: |
        Exige `X-API-Key` com uma chave mestra não vazia de `WACALLS_API_KEY`.
        Não aceita `?apiKey=`, chave de widget ou URL assinada como substituto.
        O nome do arquivo não é uma capability. O acesso é auditado e a resposta
        usa `Cache-Control: no-store`. Não há fallback público. Arquivos removidos
        pela política do arquivo de conversas retornam 404.
      security:
        - apiKeyAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: "abc123.mp3" }
      responses:
        "200":
          description: Áudio MP3
          content:
            audio/mpeg: { schema: { type: string, format: binary } }
        "401": { description: Chave mestra ausente ou inválida no header }
        "404": { description: Não encontrado }
        "503": { description: Não foi possível registrar a auditoria do acesso }
...